+++ tests/TestCase/Command/ExportCommandTest.php
@@ -116,35 +116,6 @@ class ExportCommandTest extends TestCase
     }
 
     /**
-     * Test for `execute()` method, with `send` option.
-     *
-     * @uses \DatabaseBackup\Command\ExportCommand::execute()
-     */
-    #[Test]
-    public function testExecuteSendOption(): void
-    {
-        Configure::write('DatabaseBackup.mailSender', 'sender@example.com');
-
-        $this->exec($this->command . ' --send mymail@example.com');
-        $this->assertExitSuccess();
-        $this->assertOutputRegExp('/Backup `[\w\-\/\:\\\\]+backup_[\w_]+\.sql` has been exported/');
-        $this->assertOutputRegExp('/Backup `[\w\-\/\:\\\\]+backup_[\w_]+\.sql` was sent via mail/');
-        $this->assertErrorEmpty();
-    }
-
-    /**
-     * @uses \DatabaseBackup\Command\ExportCommand::execute()
-     */
-    #[Test]
-    #[WithoutErrorHandler]
-    public function testExecuteSendOptionIsDeprecated(): void
-    {
-        $this->deprecated(function (): void {
-            $this->exec($this->command . ' --send mymail@example.com');
-        });
-    }
-
-    /**
      * Test for `execute()` method, with `timeout` option.
      *
      * @uses \DatabaseBackup\Command\ExportCommand::execute()