+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 ## 2.14 branch
 ### 2.14.0
 * passing the `$compression` argument as a string or `null` to `BackupExport::compression()` had been deprecated and has
-  been removed (backwards compatibility removed).
+  been removed (backwards compatibility removed);
+* `DeleteAllCommand`, `RotateCommand` and `SendCommand` classes had been deprecated and have been removed.
 
 ## 2.13 branch
 ### 2.13.5